# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Scrabble game API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface UserRegistration {
  /**
     * @minLength 3
     * @maxLength 20
     */
  username: string;
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface UserLogin {
  email: string;
  password: string;
}

export interface User {
  id: string;
  username: string;
  email: string;
  /** @nullable */
  avatarUrl: string | null;
  notificationTurn: boolean;
  notificationChat: boolean;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface UserPublic {
  id: string;
  username: string;
  /** @nullable */
  avatarUrl: string | null;
}

export interface ProfileUpdate {
  /**
     * @minLength 3
     * @maxLength 20
     */
  username?: string;
}

export interface PushTokenInput {
  token: string;
}

export type FriendEntryStatus = typeof FriendEntryStatus[keyof typeof FriendEntryStatus];


export const FriendEntryStatus = {
  accepted: 'accepted',
  pending: 'pending',
} as const;

export interface FriendEntry {
  friendshipId: string;
  user: UserPublic;
  status: FriendEntryStatus;
  createdAt: string;
}

export interface FriendsList {
  friends: FriendEntry[];
  pendingReceived: FriendEntry[];
  pendingSent: FriendEntry[];
}

export type FriendshipStatus = typeof FriendshipStatus[keyof typeof FriendshipStatus];


export const FriendshipStatus = {
  pending: 'pending',
  accepted: 'accepted',
  declined: 'declined',
} as const;

export interface Friendship {
  id: string;
  requesterId: string;
  addresseeId: string;
  status: FriendshipStatus;
  createdAt: string;
}

export interface FriendRequest {
  addresseeId: string;
}

export type FriendResponseAction = typeof FriendResponseAction[keyof typeof FriendResponseAction];


export const FriendResponseAction = {
  accept: 'accept',
  decline: 'decline',
} as const;

export interface FriendResponse {
  action: FriendResponseAction;
}

export interface InviteLink {
  token: string;
  expiresAt: string;
}

export interface InviteInfo {
  token: string;
  createdBy: UserPublic;
  expiresAt: string;
}

export interface BoardCell {
  /** @nullable */
  letter: string | null;
  isBlank: boolean;
  /** @nullable */
  placedBy: string | null;
}

export interface PlacedTile {
  /**
     * @minimum 0
     * @maximum 14
     */
  row: number;
  /**
     * @minimum 0
     * @maximum 14
     */
  col: number;
  letter: string;
  isBlank: boolean;
}

export interface GamePlayerState {
  userId: string;
  username: string;
  /** @nullable */
  avatarUrl: string | null;
  score: number;
  rackSize: number;
  isCurrentTurn: boolean;
}

export type GameStateStatus = typeof GameStateStatus[keyof typeof GameStateStatus];


export const GameStateStatus = {
  waiting: 'waiting',
  active: 'active',
  finished: 'finished',
} as const;

export interface GameState {
  id: string;
  status: GameStateStatus;
  board: BoardCell[][];
  players: GamePlayerState[];
  myRack: string[];
  bagSize: number;
  consecutivePasses: number;
  /** @nullable */
  turnDeadlineAt: string | null;
  /** @nullable */
  winnerId: string | null;
  createdAt: string;
  updatedAt: string;
}

export type GameSummaryStatus = typeof GameSummaryStatus[keyof typeof GameSummaryStatus];


export const GameSummaryStatus = {
  waiting: 'waiting',
  active: 'active',
  finished: 'finished',
} as const;

export interface GameSummary {
  id: string;
  status: GameSummaryStatus;
  opponent: UserPublic;
  myScore: number;
  opponentScore: number;
  isMyTurn: boolean;
  /** @nullable */
  turnDeadlineAt: string | null;
  /** @nullable */
  winnerId: string | null;
  updatedAt: string;
}

export interface GameCreate {
  opponentId: string;
}

export interface TilePlacement {
  tiles: PlacedTile[];
}

export interface TileSwap {
  letters: string[];
}

export interface MoveResult {
  success: boolean;
  gameState: GameState;
  wordsFormed: string[];
  score: number;
  /** @nullable */
  message: string | null;
}

export type GameMoveMoveType = typeof GameMoveMoveType[keyof typeof GameMoveMoveType];


export const GameMoveMoveType = {
  place: 'place',
  swap: 'swap',
  pass: 'pass',
  resign: 'resign',
} as const;

export interface GameMove {
  id: string;
  userId: string;
  username: string;
  moveType: GameMoveMoveType;
  /** @nullable */
  tilesPlaced: PlacedTile[] | null;
  wordsFormed: string[];
  score: number;
  createdAt: string;
}

export interface WordHint {
  /** @nullable */
  word: string | null;
  tiles: PlacedTile[];
  score: number;
}

export type WordStrengthStrength = typeof WordStrengthStrength[keyof typeof WordStrengthStrength];


export const WordStrengthStrength = {
  weak: 'weak',
  fair: 'fair',
  good: 'good',
  great: 'great',
  exceptional: 'exceptional',
} as const;

export interface WordStrength {
  valid: boolean;
  score: number;
  strength: WordStrengthStrength;
  words: string[];
}

export interface ChatMessage {
  id: string;
  gameId: string;
  userId: string;
  username: string;
  /** @nullable */
  avatarUrl: string | null;
  message: string;
  createdAt: string;
}

export interface ChatMessageInput {
  /**
     * @minLength 1
     * @maxLength 500
     */
  message: string;
}

export interface PlayerStats {
  userId: string;
  username: string;
  gamesPlayed: number;
  gamesWon: number;
  gamesLost: number;
  winRate: number;
  averageScore: number;
  bestScore: number;
  /** @nullable */
  bestWord: string | null;
  bestWordScore: number;
  totalPoints: number;
}

export interface NotificationSettings {
  notificationTurn: boolean;
  notificationChat: boolean;
}

export interface NotificationSettingsUpdate {
  notificationTurn?: boolean;
  notificationChat?: boolean;
}

export type SearchUsersParams = {
q: string;
};