# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Scrabble game API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRegistration"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "409":
          description: Username or email taken
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserLogin"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized

  # Users
...
  /friends:
    get:
      operationId: listFriends
      tags: [friends]
      summary: List friends and pending requests
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Friends list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FriendsList"
  /friends/request:
    post:
      operationId: sendFriendRequest
      tags: [friends]
      summary: Send a friend request
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FriendRequest"
      responses:
        "201":
          description: Request sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Friendship"
        "409":
          description: Already friends or request pending
  /friends/{friendshipId}:
    patch:
      operationId: respondFriendRequest
      tags: [friends]
      summary: Accept or decline a friend request
      security:
        - bearerAuth: []
      parameters:
        - name: friendshipId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FriendResponse"
      responses:
        "200":
          description: Updated friendship
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Friendship"
    delete:
      operationId: removeFriend
      tags: [friends]
      summary: Remove a friend or cancel request
      security:
        - bearerAuth: []
      parameters:
        - name: friendshipId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /friends/invite-link:
    post:
      operationId: createInviteLink
      tags: [friends]
      summary: Create a shareable invite link
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Invite link created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InviteLink"
  /friends/invite/{token}:
    get:
      operationId: getInviteInfo
      tags: [friends]
      summary: Get info about an invite link
      security:
        - bearerAuth: []
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Invite info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InviteInfo"
        "404":
          description: Invalid or expired invite

    post:
      operationId: acceptInviteLink
      tags: [friends]
      summary: Accept an invite link (add as friend)
      security:
        - bearerAuth: []
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Friend request sent or accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Friendship"

  # Games
...
  /games/{gameId}/chat:
    get:
      operationId: getChatMessages
      tags: [chat]
      summary: Get chat messages for a game
      security:
        - bearerAuth: []
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Chat messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatMessage"
    post:
      operationId: sendChatMessage
      tags: [chat]
      summary: Send a chat message
      security:
        - bearerAuth: []
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatMessageInput"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatMessage"

  # Stats
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success

    # Auth schemas
    UserRegistration:
      type: object
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 20
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
      required:
        - username
        - email
        - password
    UserLogin:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - token
        - user

    # User schemas
    User:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        email:
          type: string
        avatarUrl:
          type: ["string", "null"]
        notificationTurn:
          type: boolean
        notificationChat:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - username
        - email
        - avatarUrl
        - notificationTurn
        - notificationChat
        - createdAt
    UserPublic:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        avatarUrl:
          type: ["string", "null"]
      required:
        - id
        - username
        - avatarUrl
...
    FriendsList:
      type: object
      properties:
        friends:
          type: array
          items:
            $ref: "#/components/schemas/FriendEntry"
        pendingReceived:
          type: array
          items:
            $ref: "#/components/schemas/FriendEntry"
        pendingSent:
          type: array
          items:
            $ref: "#/components/schemas/FriendEntry"
      required:
        - friends
        - pendingReceived
        - pendingSent
    FriendEntry:
      type: object
      properties:
        friendshipId:
          type: string
        user:
          $ref: "#/components/schemas/UserPublic"
        status:
          type: string
          enum: [accepted, pending]
        createdAt:
          type: string
          format: date-time
      required:
        - friendshipId
        - user
        - status
        - createdAt
    Friendship:
      type: object
      properties:
        id:
          type: string
        requesterId:
          type: string
        addresseeId:
          type: string
        status:
          type: string
          enum: [pending, accepted, declined]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - requesterId
        - addresseeId
        - status
        - createdAt
    FriendRequest:
      type: object
      properties:
        addresseeId:
          type: string
      required:
        - addresseeId
    FriendResponse:
      type: object
      properties:
        action:
          type: string
          enum: [accept, decline]
      required:
        - action
    InviteLink:
      type: object
      properties:
        token:
          type: string
        expiresAt:
          type: string
          format: date-time
      required:
        - token
        - expiresAt
    InviteInfo:
      type: object
      properties:
        token:
          type: string
        createdBy:
          $ref: "#/components/schemas/UserPublic"
        expiresAt:
          type: string
          format: date-time
      required:
        - token
        - createdBy
        - expiresAt

    # Game schemas
...
    ChatMessage:
      type: object
      properties:
        id:
          type: string
        gameId:
          type: string
        userId:
          type: string
        username:
          type: string
        avatarUrl:
          type: ["string", "null"]
        message:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - gameId
        - userId
        - username
        - avatarUrl
        - message
        - createdAt
    ChatMessageInput:
      type: object
      properties:
        message:
          type: string
          minLength: 1
          maxLength: 500
      required:
        - message

    # Stats schemas
...