# lib/api-zod/src/generated/types/friendEntry.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Scrabble game API
 * OpenAPI spec version: 0.1.0
 */
import type { FriendEntryStatus } from './friendEntryStatus';
import type { UserPublic } from './userPublic';

export interface FriendEntry {
  friendshipId: string;
  user: UserPublic;
  status: FriendEntryStatus;
  createdAt: Date;
}